Added timestamp to avoid time drifting
When hibernating or putting your computer to sleep the countdown will
temporarily stop and then resume when computer is awaken. I want to
ensure that the elapsed time isn't beyond the time specified in the
countdown value.

For example if countdown value is set to 1200 seconds and user
hibernates computer after 10 minutes, then resumes the session 30
minutes after that, then the countdown value is automatically set to 0,
0 is emitted and countdown is stopped.

If a new countdown value is set then a new timestamp is set as well.